[BUGFIX] Content type set for message is not respected
When a custom content type, like text/plain for text-only mails, is set,
the message is sent as text/html nonetheless, because the charset
setting from the SwiftMailer class is not respected.

This fixes an earlier patchset which somehow mixed "content type" and
"charset" and also made the change at a too early point.